CCommandLineParamsAdapterComp fixed
git-svn-id: http://b035a0a.online-server.cloud/svn/ITDevelopment@5267 21205d9f-8918-2240-a3d5-c39064a2c852

+++ Include/imtbase/CCommandLineParamsAdapterComp.cpp
@@ -83,7 +83,7 @@ void CCommandLineParamsAdapterComp::OnUpdate(const istd::IChangeable::ChangeSet&
 					case PT_DOUBLE:
 						{
 							bool isOk;
-							int value = optionValue.toDouble(&isOk);
+							double value = optionValue.toDouble(&isOk);
 							if (!isOk){
 								ResetData();
 								return;
